refactor(footer): to typescript
